+++ src/parser/block/extension/fig.ts
@@ -8,6 +8,7 @@ import { segment as seg_math } from '../mathblock';
 import { segment as seg_table } from './table';
 import { segment as seg_blockquote } from '../blockquote';
 import { segment as seg_placeholder } from './placeholder';
+import { media, shortmedia } from '../../inline';
 
 import FigParser = ExtensionParser.FigParser;
 
@@ -25,10 +26,21 @@ export const segment: FigParser.SegmentParser = block(validate(['[$', '$'],
   ])));
 
 export const fig: FigParser = block(rewrite(segment, verify(convert(
-  source => {
+  (source, context) => {
     const fence = (/^[^\n]*\n!?>+\s/.test(source) && source.match(/^~{3,}(?=[^\S\n]*$)/mg) || [])
       .reduce((max, fence) => fence > max ? fence : max, '~~') + '~';
-    return `${fence}figure ${source}\n\n${fence}`;
+    return parser({ source, context })
+      ? `${fence}figure ${source.replace(/^(.+\n.+\n)([\S\s]+?)\n?$/, '$1\n$2')}\n${fence}`
+      : `${fence}figure ${source}\n\n${fence}`;
   },
   union([figure])),
   ([el]) => el.tagName === 'FIGURE')));
+
+const parser = sequence([
+  line(close(seg_label, /^(?=\s).*\n/)),
+  line(union([
+    media,
+    shortmedia,
+  ])),
+  some(contentline),
+]);